track-privacy-head refactor to BlockAddedObserver
