+++ ethereum/core/src/main/java/org/hyperledger/besu/ethereum/MainnetBlockValidator.java
@@ -16,34 +16,19 @@ package org.hyperledger.besu.ethereum;
 
 import static org.apache.logging.log4j.LogManager.getLogger;
 
-import org.hyperledger.besu.enclave.Enclave;
-import org.hyperledger.besu.enclave.EnclaveException;
-import org.hyperledger.besu.enclave.types.ReceiveRequest;
-import org.hyperledger.besu.enclave.types.ReceiveResponse;
 import org.hyperledger.besu.ethereum.chain.MutableBlockchain;
-import org.hyperledger.besu.ethereum.core.Address;
 import org.hyperledger.besu.ethereum.core.Block;
 import org.hyperledger.besu.ethereum.core.BlockHeader;
-import org.hyperledger.besu.ethereum.core.Hash;
 import org.hyperledger.besu.ethereum.core.MutableWorldState;
-import org.hyperledger.besu.ethereum.core.PrivacyParameters;
 import org.hyperledger.besu.ethereum.core.TransactionReceipt;
 import org.hyperledger.besu.ethereum.mainnet.BlockBodyValidator;
 import org.hyperledger.besu.ethereum.mainnet.BlockHeaderValidator;
 import org.hyperledger.besu.ethereum.mainnet.BlockProcessor;
 import org.hyperledger.besu.ethereum.mainnet.HeaderValidationMode;
-import org.hyperledger.besu.ethereum.privacy.storage.PrivateGroupIdToLatestBlockWithTransactionMap;
-import org.hyperledger.besu.ethereum.privacy.storage.PrivateStateStorage;
-import org.hyperledger.besu.util.bytes.Bytes32;
-import org.hyperledger.besu.util.bytes.BytesValue;
-import org.hyperledger.besu.util.bytes.BytesValues;
 
-import java.util.Collections;
-import java.util.HashMap;
 import java.util.List;
 import java.util.Optional;
 
-import com.google.common.collect.Maps;
 import org.apache.logging.log4j.Logger;
 
 public class MainnetBlockValidator<C> implements BlockValidator<C> {
@@ -55,30 +40,14 @@ public class MainnetBlockValidator<C> implements BlockValidator<C> {
   private final BlockBodyValidator<C> blockBodyValidator;
 
   private final BlockProcessor blockProcessor;
-  private PrivacyParameters privacyParameters;
-
-  private PrivateStateStorage privateStateStorage;
-  private Address address;
-  private Enclave enclave;
 
   public MainnetBlockValidator(
       final BlockHeaderValidator<C> blockHeaderValidator,
       final BlockBodyValidator<C> blockBodyValidator,
-      final BlockProcessor blockProcessor,
-      final PrivacyParameters privacyParameters) {
+      final BlockProcessor blockProcessor) {
     this.blockHeaderValidator = blockHeaderValidator;
     this.blockBodyValidator = blockBodyValidator;
     this.blockProcessor = blockProcessor;
-    this.privacyParameters = privacyParameters;
-    setUpForPrivacy(privacyParameters);
-  }
-
-  private void setUpForPrivacy(final PrivacyParameters privacyParameters) {
-    if (privacyParameters.isEnabled()) {
-      privateStateStorage = privacyParameters.getPrivateStateStorage();
-      address = Address.privacyPrecompiled(privacyParameters.getPrivacyAddress());
-      enclave = new Enclave(privacyParameters.getEnclaveUri());
-    }
   }
 
   @Override
@@ -119,38 +88,6 @@ public class MainnetBlockValidator<C> implements BlockValidator<C> {
     final BlockProcessor.Result result = blockProcessor.processBlock(blockchain, worldState, block);
     if (!result.isSuccessful()) {
       return Optional.empty();
-    }
-
-    if (privacyParameters.isEnabled()) {
-      // get all groups
-      final HashMap<Bytes32, Hash> privacyGroupToLatestBlockWithTransactionMap =
-          Maps.newHashMap(
-              privateStateStorage
-                  .getPrivacyGroupToLatestBlockWithTransactionMap(block.getHeader().getParentHash())
-                  .map(PrivateGroupIdToLatestBlockWithTransactionMap::getMap)
-                  .orElse(Collections.emptyMap()));
-      block.getBody().getTransactions().stream()
-          .filter(t -> t.getTo().isPresent() && t.getTo().get().equals(address))
-          .forEach(
-              t -> {
-                final String key = BytesValues.asBase64String(t.getData().get());
-                try {
-                  final ReceiveResponse receiveResponse = enclave.receive(new ReceiveRequest(key));
-                  final BytesValue privacyGroupId =
-                      BytesValues.fromBase64(receiveResponse.getPrivacyGroupId());
-                  privacyGroupToLatestBlockWithTransactionMap.put(
-                      Bytes32.wrap(privacyGroupId), block.getHash());
-                } catch (final EnclaveException e) {
-                  LOG.warn("Enclave does not have private transaction with key {}.", key, e);
-                }
-              });
-      privateStateStorage
-          .updater()
-          .putPrivacyGroupToLatestBlockWithTransactionMap(
-              block.getHash(),
-              new PrivateGroupIdToLatestBlockWithTransactionMap(
-                  privacyGroupToLatestBlockWithTransactionMap))
-          .commit();
     }
 
     final List<TransactionReceipt> receipts = result.getReceipts();
