+++ ethereum/retesteth/src/main/java/org/hyperledger/besu/ethereum/retesteth/NoRewardProtocolScheduleWrapper.java
@@ -17,7 +17,6 @@ package org.hyperledger.besu.ethereum.retesteth;
 import org.hyperledger.besu.ethereum.BlockValidator;
 import org.hyperledger.besu.ethereum.MainnetBlockValidator;
 import org.hyperledger.besu.ethereum.core.BlockImporter;
-import org.hyperledger.besu.ethereum.core.PrivacyParameters;
 import org.hyperledger.besu.ethereum.core.TransactionFilter;
 import org.hyperledger.besu.ethereum.core.Wei;
 import org.hyperledger.besu.ethereum.mainnet.BlockProcessor;
@@ -51,8 +50,7 @@ public class NoRewardProtocolScheduleWrapper<C> implements ProtocolSchedule<C> {
         new MainnetBlockValidator<>(
             original.getBlockHeaderValidator(),
             original.getBlockBodyValidator(),
-            noRewardBlockProcessor,
-            PrivacyParameters.DEFAULT);
+            noRewardBlockProcessor);
     final BlockImporter<C> noRewardBlockImporter =
         new MainnetBlockImporter<>(noRewardBlockValidator);
     return new ProtocolSpec<>(
